Alter DistributionException to use %s not %g.
Many distributions have integer parameters. These cannot be formatted
using %g. The alternative is to cast them all to a double. The output
would have a decimal point and trailing zeros:

1 => 1.0000

+++ commons-statistics-distribution/src/main/java/org/apache/commons/statistics/distribution/AbstractDiscreteDistribution.java
@@ -37,7 +37,7 @@ abstract class AbstractDiscreteDistribution
                               int x1) {
         if (x1 < x0) {
             throw new DistributionException(DistributionException.TOO_SMALL,
-                                            (double) x1, (double) x0);
+                                            x1, x0);
         }
         return cumulativeProbability(x1) - cumulativeProbability(x0);
     }
